Simplify file structure around PermArgs and EvalArgs (#14)
