--- triton-vm/src/arguments.rs
@@ -1,2 +1,0 @@
-pub mod evaluation_argument;
-pub mod permutation_argument;
